Fixed the final few bugs in handling system, really happy with it now. Redid the car rendering, making it more efficient and also implemented color schemes. Started phasing out the dependency on glm in favor of SFML transformations.

+++ src/resources/color_scheme.hpp
@@ -17,7 +17,7 @@ namespace ts
   {
     struct ColorScheme
     {
-      std::uint16_t scheme_id = 0;
+      std::uint32_t scheme_id = 0;
       std::array<Colorb, 3> colors;
     };
   }
